feat: 무한 스크롤 기능 및 소수점 JSONParsing 에러 해결

+++ SwiftUI-open-market/SwiftUI-open-market/Model/Product.swift
@@ -12,7 +12,7 @@ struct Product: Codable {
     let vendorName, name, pageDescription: String
     let thumbnail: String
     let currency: String
-    let price, bargainPrice, discountedPrice, stock: Int
+    let price, bargainPrice, discountedPrice, stock: Float
     let createdAt, issuedAt: String
 
     enum CodingKeys: String, CodingKey {
